Rename client (for local fs) to local_client
